+++ src/mb_optimizer/__init__.py
@@ -1,3 +1,3 @@
 """MB CF Optimizer."""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"